Move tasks to notice_comment module.
